+++ app/Http/Controllers/PostsController.php
@@ -3,8 +3,8 @@
 namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
-use App\Post;
-use App\Comment;
+use App\{Post, Comment, Account};
+use Laratrust;
 
 class PostsController extends Controller
 {
@@ -16,13 +16,23 @@ class PostsController extends Controller
 
     public function index()
     {
-        $posts = Post::latest()->simplePaginate(10);
+        if ($this->isAdminRequest())
+        {
+            return $this->adminIndex();
+        }
+        else
+        {
+            $posts = Post::latest()->simplePaginate(10);
 
-        return view('posts.index', compact('posts'));
+            return view('posts.index', compact('posts'));
+        }
     }
 
     public function show(Post $post)
     {
+        if (auth()->check() && !Laratrust::can('view-post'))
+            return abort(403);
+
         $comments = Comment::with('account')->wherePostId($post->id)->simplePaginate(10);
 
         return view('posts.show', compact('post', 'comments'));
@@ -30,24 +40,46 @@ class PostsController extends Controller
 
     public function edit(Post $post)
     {
-        return view('posts.edit', compact('post'));
+        if (!Laratrust::can('edit-post') && !Laratrust::canAndOwns('update-own-post', $post))
+            return abort(403);
+
+        return $this->isAdminRequest() ? view('admin.posts.edit', compact('post')) : view('posts.edit', compact('post'));
     }
 
     public function update(Post $post)
     {
+        if (!Laratrust::can('edit-post') && !Laratrust::canAndOwns('update-own-post', $post))
+            return abort(403);
+
         $this->postValidation();
-        $post->update(request(['title', 'content']));
+        $this->validate(request(), [
+            'account_id' => 'sometimes|required|numeric'
+        ]);
+
+        $post->title = request()->title;
+        $post->content = request()->content;
+
+        if ( isset(request()->account_id) )
+            $post->account()->associate(Account::findOrFail(request()->account_id));
+
+        $post->save();
 
         return redirect()->route('posts.show', ['id' => $post->id]);
     }
 
     public function create()
     {
-        return view('posts.create');
+        if (!Laratrust::can('create-post'))
+            return abort(403);
+
+        return $this->isAdminRequest() ? view('admin.posts.create') : view('posts.create');
     }
 
     public function store()
     {
+        if (!Laratrust::can('create-post'))
+          return abort(403);
+
         $this->postValidation();
 
         Post::create([
@@ -61,9 +93,26 @@ class PostsController extends Controller
 
     public function destroy(Post $post)
     {
+        if (!Laratrust::can('delete-post') && !Laratrust::canAndOwns('delete-own-post', $post))
+          return abort(403);
+
         $post->delete();
 
         return redirect('/');
+    }
+
+    private function adminIndex()
+    {
+        $posts = Post::with('account');
+
+        if (request()->keywords)
+            $posts->where('content', 'LIKE', '%'.request()->keywords.'%');
+
+        $posts = $posts->paginate(10);
+
+        request()->flashOnly(['keywords']);
+
+        return view('admin.posts.index', compact('posts'));
     }
 
     private function postValidation()
